requirements.txt
# Core Dependencies
matplotlib>=3.8.0
Pillow>=10.3.0
requests>=2.31.0
ttkbootstrap>=0.5.3
pytest>=7.3.0
pytest-selenium>=4.0.0
pytest-html>=4.1.0

# Selenium & Browser Automation
selenium>=4.15.0
undetected-chromedriver>=3.5.4
pyautogui>=0.9.54

# Data Processing
pandas>=2.1.0
beautifulsoup4>=4.12.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.1
reportlab>=4.0.0
fpdf>=1.7.2

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.35.0

# AI Providers
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0

# GUI
tkintertest>=1.0.0

# Utilities
python-dateutil>=2.8.0
psutil>=5.9.0
